init stock check

+++ src/constants/schema.ts
@@ -80,8 +80,8 @@ export const createImportReceiptSchema = z.object({
 export const stockCheckItemSchema = z.object({
   productId: z.string().uuid('Vui lòng chọn sản phẩm'),
   unitId: z.string().uuid('Vui lòng chọn đơn vị'),
-  quantityInStock: z.coerce.number().int().min(0, 'Tồn kho phải >= 0'),
-  quantityActual: z.coerce.number().int().min(0, 'Thực tế phải >= 0'),
+  quantityInStock: z.coerce.number().min(0, 'Tồn kho phải >= 0'),
+  quantityActual: z.coerce.number().min(0, 'Thực tế phải >= 0'),
 })
 
 export const stockCheckSchema = z.object({
